fix(client): keep Amp auth sanitizer behind the provider seam
Move URL/query stripping and redactErrorPreview composition into the Amp
provider. Shared formatAuthHint stays catalog-driven with no Runtime import.
