+++ packages/client/src/__tests__/auth-error-hint.test.ts
@@ -6,7 +6,6 @@ import {
   isCodexAuthError,
   isGrokAuthError,
   isOpenCodeAuthError,
-  sanitizeAmpAuthFailureText,
 } from "../providers/handlers/auth-error-hint.js";
 
 /**
@@ -188,34 +187,6 @@ describe("formatAuthHint", () => {
     expect(hint).toContain("`amp login`");
     expect(hint).toContain("Amp");
     expect(hint).toContain("not First Tree's");
-  });
-
-  it("discards Amp login URLs and generic credential shapes from the chat hint", () => {
-    const official = [
-      "No API key found. Starting login flow... AMP_API_KEY=qa-amp-key-placeholder Authorization: Bearer qa-bearer-placeholder sk-ant-abcdefghijklmnopqrstuvwxyz012345",
-      "If your browser does not open automatically, visit:",
-      "",
-      "https://ampcode.com/auth/cli-login?authToken=qa-one-time-placeholder&state=qa-state-placeholder",
-      "",
-      "When prompted, paste your code here:",
-    ].join("\n");
-    const sanitized = sanitizeAmpAuthFailureText(official);
-    expect(sanitized).not.toMatch(/https?:\/\//i);
-    expect(sanitized).not.toMatch(/authToken=|state=|[?&]code=/i);
-    expect(sanitized).not.toContain("qa-amp-key-placeholder");
-    expect(sanitized).not.toContain("qa-bearer-placeholder");
-    expect(sanitized).not.toContain("sk-ant-abcdefghijklmnopqrstuvwxyz012345");
-    expect(sanitized).toContain("No API key found");
-    const hint = formatAuthHint("amp", official);
-    expect(hint).toContain("`amp login`");
-    expect(hint).toContain("No API key found");
-    expect(hint).not.toMatch(/https?:\/\//i);
-    expect(hint).not.toMatch(/authToken=|state=|[?&]code=/i);
-    expect(hint).not.toContain("qa-one-time-placeholder");
-    expect(hint).not.toContain("qa-amp-key-placeholder");
-    expect(hint).not.toContain("qa-bearer-placeholder");
-    expect(hint).not.toContain("sk-ant-abcdefghijklmnopqrstuvwxyz012345");
-    expect(hint).not.toContain("paste your code");
   });
 
   it("keeps OpenCode authentication host-local and points at the provider-owned login", () => {
