+++ packages/client/src/providers/handlers/auth-error-hint.ts
@@ -18,7 +18,6 @@ import {
   runtimeProviderAuthOwnerLabel,
   runtimeProviderChatAuthLoginPhrase,
 } from "@first-tree/shared";
-import { redactErrorPreview } from "../../runtime/redact-error-preview.js";
 
 type Runtime = Exclude<RuntimeProvider, "claude-code-tui">;
 
@@ -171,29 +170,6 @@ export function isClaudeAuthError(code: string | undefined): boolean {
 }
 
 /**
- * Drop complete login URLs and one-time query material (authToken/code/state)
- * before any provider auth text becomes chat-visible or durable. Amp's official
- * absent-key path prints `https://ampcode.com/auth/cli-login?authToken=…` on
- * stdout; First Tree must not retain that authorization material.
- */
-export function discardProviderLoginAuthorizationMaterial(text: string): string {
-  return text
-    .replace(/https?:\/\/[^\s<>"']+/gi, "")
-    .replace(/(?:^|[?&\s/#])(?:authToken|auth_token|token|code|state|auth_code)=[^\s&"'<>]+/gi, " ")
-    .replace(/\bAMP_API_KEY\s*[:=]\s*[^\s&"'<>]+/gi, "AMP_API_KEY=[REDACTED]")
-    .replace(/when prompted,\s*paste your code here:?\s*/gi, "")
-    .replace(/[ \t]+\n/g, "\n")
-    .replace(/\n{3,}/g, "\n\n")
-    .replace(/[ \t]{2,}/g, " ")
-    .trim();
-}
-
-/** Amp chat-safe auth text: drop login URLs/query, then generic credential redaction. */
-export function sanitizeAmpAuthFailureText(text: string): string {
-  return redactErrorPreview(discardProviderLoginAuthorizationMaterial(text), Number.POSITIVE_INFINITY);
-}
-
-/**
  * Build the chat-timeline message we want the user to see when an auth
  * failure is detected. Includes the raw SDK error verbatim so the user can
  * paste it into a support thread without losing detail. The hint is short
@@ -208,8 +184,7 @@ export function formatAuthHint(runtime: Runtime, originalMessage: string): strin
   // Cap the appended raw message so an upstream stack-trace envelope (codex
   // wraps its `event.error.message` in surprising ways) doesn't bloat the
   // hint into a wall of text on the chat timeline.
-  const source = runtime === "amp" ? sanitizeAmpAuthFailureText(originalMessage) : originalMessage;
-  const trimmed = source.trim().slice(0, ORIGINAL_MESSAGE_CAP);
+  const trimmed = originalMessage.trim().slice(0, ORIGINAL_MESSAGE_CAP);
   const original = trimmed.length > 0 ? trimmed : "(no message from SDK)";
   return (
     `${runtime} auth on this machine looks broken or expired. ` +
